backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
qdrant-client
elasticsearch
sentence-transformers
pandas
